+++ api/src/main/java/com/homeaway/datapullclient/input/ClusterProperties.java
@@ -103,6 +103,8 @@ public class ClusterProperties {
     @JsonAlias({"Tags", "tags"})
     private Map<String, String> tags = new HashMap<String, String>();
 
+    @JsonProperty("bootstrapactionstring")
+    private String bootstrapactionstring;
 
     private String env;
 
@@ -128,6 +130,7 @@ public class ClusterProperties {
                 ", portfolio='" + portfolio + '\'' +
                 ", product='" + product + '\'' +
                 ", team='" + team + '\'' +
+                ", bootstrapactionstring='" + bootstrapactionstring + '\'' +
                 ", env='" + env + '\'' +
                 ", costCenter='" + costCenter + '\'' +
                 '}';